+++ keyboards/kc60/keymaps/noroadsleft/readme_ch5.md
@@ -36,34 +36,34 @@ Tapping `Esc` exits the Macro layer, if the macro used doesn't do it automatical
 
 ### Macros
 
-[![lavak3DED](https://static-cdn.jtvnw.net/emoticons/v1/821796/1.0)](../../../../users/noroadsleft/noroadsleft.h#L76-L81)  
+[![lavak3DED](https://static-cdn.jtvnw.net/emoticons/v1/821796/1.0)](./keymap.c#L122-L127)  
 Inputs: `lavak3DED `  
 Twitch emote for [a streamer I watch a lot](https://www.twitch.tv/lavak3_).
 
-[G_PUSH](../../../../users/noroadsleft/noroadsleft.h#L82-L87)  
+[G_PUSH](./keymap.c#L128-L133)  
 Inputs: `git push origin `  
 Everything from here down is related to Git or GitHub.
 
-[G_FTCH](../../../../users/noroadsleft/noroadsleft.h#L88-L93)  
+[G_FTCH](./keymap.c#L134-L139)  
 Inputs: `git fetch upstream`
 
-[G_COMM](../../../../users/noroadsleft/noroadsleft.h#L94-L100)  
+[G_COMM](./keymap.c#L140-L146)  
 Inputs: `git commit -m ""` `[Left]`  
 Readies a `git commit` command, moves the cursor between the quotation marks, then disables the Macro layer.
 
-[G_RST](../../../../users/noroadsleft/noroadsleft.h#L101-L107)  
+[G_RST](./keymap.c#L147-L153)  
 Inputs: `git histt -n 10`, Enter, `git reset --soft `  
 Runs a [git alias](./readme_git.md) that shows my last ten commits, then readies a `git reset --soft`. For when I commit something too soon. Disables the Macro layer when complete.
 
-[G_C10R](../../../../users/noroadsleft/noroadsleft.h#L108-L114)  
+[G_C10R](./keymap.c#L154-L160)  
 Inputs: `cf/`  
 A branch prefix I use for my current work in QMK. Disables the Macro layer when finished.
 
-[G_BRCH](../../../../users/noroadsleft/noroadsleft.h#L115-L121)  
+[G_BRCH](./keymap.c#L161-L167)  
 Inputs: `$(git branch-name)`  
 A [git alias](./readme_git.md) that returns the name of the current branch. Disables the Macro layer when finished.
 
-[SIGNA](../../../../users/noroadsleft/noroadsleft.h#L107-L110)  
+[SIGNA](./keymap.c#L168-L174)  
 Inputs: `\- @noroadsleft` `[Enter]`  
 Sometimes on GitHub, I sign my comments. Types my GitHub name in Markdown syntax, and then taps the `Enter` key. Disables the Macro layer when finished.
